art_lesson_planner_agent/requirements.txt
crewai>=0.30.0
anthropic>=0.39.0
manim>=0.18.0
pydantic>=2.0.0
asyncio
python-dotenv
numpy
matplotlib
typing-extensions